+++ src/jtl/Connector/Shopware/Mapper/ProductPrice.php
@@ -60,14 +60,12 @@ class ProductPrice extends DataMapper
      * @param \Shopware\Models\Article\Article $productSW
      * @param \Shopware\Models\Article\Detail $detailSW
      * @param float $recommendedRetailPrice
-     * @param float $purchasePrice
      */
     public static function buildCollection(
         array $productPrices,
         ArticleSW &$productSW = null,
         DetailSW &$detailSW = null,
-        $recommendedRetailPrice = null,
-        $purchasePrice = null
+        $recommendedRetailPrice = null
     ) {
         // Price
         $collection = [];
@@ -167,14 +165,6 @@ class ProductPrice extends DataMapper
             );
         }
 
-        // Find baseprice
-        if ($productSW->getId() > 0 && $detailSW->getId() > 0 && is_null($purchasePrice)) {
-            $purchasePrice = Shopware()->Db()->fetchOne(
-                'SELECT if(baseprice, baseprice, 0.0) FROM s_articles_prices WHERE articleID = ? AND articledetailsID = ? AND `from` = 1',
-                array($productSW->getId(), $detailSW->getId())
-            );
-        }
-
         $sql = "DELETE FROM s_articles_prices WHERE articleID = ? AND articledetailsID = ?";
         Shopware()->Db()->query($sql, array($productSW->getId(), $detailSW->getId()));
 
@@ -228,7 +218,6 @@ class ProductPrice extends DataMapper
                         ->setCustomerGroup($customerGroupSW)
                         ->setFrom($quantity)
                         ->setPrice($priceItem->getNetPrice())
-                        //->setBasePrice($purchasePrice)
                         ->setPseudoPrice($recommendedRetailPrice)
                         ->setDetail($detailSW);